% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maltq_to_fb.R
\name{maltq_to_fb}
\alias{maltq_to_fb}
\title{Convert a malt quality dataset to Field Book}
\usage{
maltq_to_fb(maltq.df, lab.to.plot, field.book)
}
\arguments{
\item{maltq.df}{A \code{data.frame} of raw malting quality data compiled
using the \code{\link{extract.maltq}} function with \code{tidy = FALSE}. The first
two columns should be the lab number (LabNo.), and the entry, (VarietyorSelection),
the last column should be the batch (Batch), and the remaining columns should
be the malt quality traits.}

\item{lab.to.plot}{A \code{data.frame} with lab number (LabNo.) and the
corresponding plot number of the trial from which grain was sampled for malting
analysis.}

\item{field.book}{A \code{data.frame} of the blank field.book table listing plot_id,
plot, trial, line_name, row, column, entry, rep, and notes.}
}
\value{
A modified field.book \code{data.frame} with the raw data from malting
quality.
}
\description{
Takes a data.frame of malt quality data and converts it
to a Field Book table format. This allows for easy database upload and for
spatial adjustment
}

