% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_malt_quality.R
\name{read_maltq}
\alias{read_maltq}
\title{Read in malt quality data}
\usage{
read_maltq(files, tidy = FALSE)
}
\arguments{
\item{files}{A \code{character} vector of filenames or filepaths of the malt
quality Excel files to read in.}

\item{tidy}{A \code{logical} indicating whether the output should be in tidy
format. See \code{Details} or \code{\link[tidyr]{gather}}}
}
\value{
A \code{list} with the following elements:

$data A \code{data.frame} of malting quality values for experimental entries

$stats A \code{data.frame} of summary statistics from each of the batches

$checks A \code{data.frame} of malting quality values for the laboratory checks
}
\description{
Loads malt quality data typically received in the University of Minnesota
Barley Breeding Group. This data often comes from the USDA Cereal Laboratory
in Madison, Wisconsin. Data from this lab is usually formatted consistently,
permitting use of a common function.
}

