% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/t3_traits.R
\docType{data}
\name{t3_traits}
\alias{t3_traits}
\title{T3 Trait Descriptions}
\format{A data frame with 138 rows and 8 columns:
\describe{
   \item{Trait}{The name used by T3 for the trait}
   \item{Nickname}{A nickname for the trait}
   \item{Ontology}{}
   \item{Description}{}
   \item{Min}{Minimum value of the trait}
   \item{Max}{Maximum value of the trait}
   \item{Unit}{Unit of the trait}
   \item{Unit.info}{}
}}
\source{
\url{https://t3sandbox.org/t3/sandbox/barley/traits.php}
}
\usage{
t3_traits
}
\description{
A \code{data.frame} of trait names and descriptions provided by the T3
database.
}
\keyword{datasets}
